src/pkg/html/testdata/webkit/plain-text-unsafe.dat

#data
FOO&#x000D;ZOO
#errors
#document
| <html>
|   <head>
|   <body>
|     "FOOZOO"
